requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.9
Pillow>=10.0.0
google-genai>=1.0.0
python-dotenv>=1.0.0
